+++ package.json
@@ -3,6 +3,16 @@
   "version": "7.2.0-alpha.1",
   "description": "A JavaScript framework for creating ambitious web applications",
   "type": "module",
+  "sideEffects": [
+    "./dist/dev/**",
+    "./dist/prod/packages/@ember/-internals/glimmer/lib/register-classic-helper.js",
+    "./dist/prod/packages/@ember/-internals/glimmer/lib/register-curly-component.js",
+    "./dist/prod/packages/@ember/-internals/glimmer/lib/syntax/register-routing-keywords.js",
+    "./dist/prod/packages/@glimmer/runtime/lib/debug-render-tree-register.js",
+    "./dist/prod/packages/@ember/-internals/runtime/lib/ext/rsvp.js",
+    "./dist/prod/packages/@glimmer/validator/index.js",
+    "./dist/prod/packages/ember-testing/index.js"
+  ],
   "keywords": [
     "ember-addon"
   ],
@@ -182,6 +192,9 @@
       "@ember/-internals/environment/index.js": "ember-source/@ember/-internals/environment/index.js",
       "@ember/-internals/error-handling/index.js": "ember-source/@ember/-internals/error-handling/index.js",
       "@ember/-internals/glimmer/index.js": "ember-source/@ember/-internals/glimmer/index.js",
+      "@ember/-internals/glimmer/lib/register-classic-helper.js": "ember-source/@ember/-internals/glimmer/lib/register-classic-helper.js",
+      "@ember/-internals/glimmer/lib/register-curly-component.js": "ember-source/@ember/-internals/glimmer/lib/register-curly-component.js",
+      "@ember/-internals/glimmer/lib/syntax/register-routing-keywords.js": "ember-source/@ember/-internals/glimmer/lib/syntax/register-routing-keywords.js",
       "@ember/-internals/meta/index.js": "ember-source/@ember/-internals/meta/index.js",
       "@ember/-internals/meta/lib/meta.js": "ember-source/@ember/-internals/meta/lib/meta.js",
       "@ember/-internals/metal/index.js": "ember-source/@ember/-internals/metal/index.js",
@@ -243,6 +256,7 @@
       "@ember/enumerable/mutable.js": "ember-source/@ember/enumerable/mutable.js",
       "@ember/helper/index.js": "ember-source/@ember/helper/index.js",
       "@ember/instrumentation/index.js": "ember-source/@ember/instrumentation/index.js",
+      "@ember/instrumentation/lib/internal-instrument.js": "ember-source/@ember/instrumentation/lib/internal-instrument.js",
       "@ember/modifier/index.js": "ember-source/@ember/modifier/index.js",
       "@ember/modifier/on.js": "ember-source/@ember/modifier/on.js",
       "@ember/object/-internals.js": "ember-source/@ember/object/-internals.js",
@@ -334,6 +348,7 @@
       "@glimmer/program/index.js": "ember-source/@glimmer/program/index.js",
       "@glimmer/reference/index.js": "ember-source/@glimmer/reference/index.js",
       "@glimmer/runtime/index.js": "ember-source/@glimmer/runtime/index.js",
+      "@glimmer/runtime/lib/debug-render-tree-register.js": "ember-source/@glimmer/runtime/lib/debug-render-tree-register.js",
       "@glimmer/tracking/index.js": "ember-source/@glimmer/tracking/index.js",
       "@glimmer/tracking/primitives/cache/index.js": "ember-source/@glimmer/tracking/primitives/cache/index.js",
       "@glimmer/util/index.js": "ember-source/@glimmer/util/index.js",
