+++ packages/@ember/-internals/glimmer/lib/component-managers/outlet.ts
@@ -2,7 +2,7 @@ import type { InternalOwner } from '@ember/-internals/owner';
 import type { Nullable } from '@ember/-internals/utility-types';
 import { assert } from '@ember/debug';
 import type EngineInstance from '@ember/engine/instance';
-import { _instrumentStart } from '@ember/instrumentation';
+import { _instrumentStart } from '@ember/instrumentation/lib/internal-instrument';
 import { precompileTemplate } from '@ember/template-compilation';
 import type {
   CompilableProgram,
@@ -21,7 +21,7 @@ import { UNDEFINED_REFERENCE, valueForRef } from '@glimmer/reference/lib/referen
 import { EMPTY_ARGS } from '@glimmer/runtime/lib/vm/arguments';
 import { unwrapTemplate } from './unwrap-template';
 
-import type { DynamicScope } from '../renderer';
+import type { DynamicScope } from '../classic-renderer';
 import type { OutletState } from '../utils/outlet';
 import type OutletView from '../views/outlet';
