+++ packages/@ember/debug/lib/capture-render-tree.ts
@@ -1,4 +1,4 @@
-import type { Renderer } from '@ember/-internals/glimmer/lib/renderer';
+import type { Renderer } from '@ember/-internals/glimmer/lib/classic-renderer';
 import type Owner from '@ember/owner';
 import type { CapturedRenderNode } from '@glimmer/interfaces';
