--- packages/@glimmer/runtime/lib/bootstrap.ts
@@ -1,8 +1,0 @@
-import './opcodes';
-import './compiled/opcodes/expressions';
-import './compiled/opcodes/component';
-import './compiled/opcodes/content';
-import './compiled/opcodes/debugger';
-import './compiled/opcodes/dom';
-import './compiled/opcodes/vm';
-import './compiled/opcodes/lists';
