+++ packages/@glimmer/runtime/lib/compiled/opcodes/debugger.ts
@@ -7,7 +7,7 @@ import { VM_DEBUGGER_OP } from '@glimmer/constants/lib/syscall-ops';
 import { unwrap } from '@glimmer/debug-util/lib/platform-utils';
 import { childRefFor, valueForRef } from '@glimmer/reference/lib/reference';
 
-import { APPEND_OPCODES } from '../../opcodes';
+import type { AppendOpcodes } from '../../opcodes';
 
 export type DebugGet = (path: string) => unknown;
 
@@ -67,8 +67,10 @@ class ScopeInspector {
   }
 }
 
-APPEND_OPCODES.add(VM_DEBUGGER_OP, (vm, { op1: _debugInfo }) => {
-  let debuggerInfo = vm.constants.getValue<DebuggerInfo>(decodeHandle(_debugInfo));
-  let inspector = new ScopeInspector(vm.scope(), debuggerInfo);
-  callback(valueForRef(vm.getSelf()), (path) => valueForRef(inspector.get(path)));
-});
+export function defineDebuggerOpcodes(APPEND_OPCODES: AppendOpcodes): void {
+  APPEND_OPCODES.add(VM_DEBUGGER_OP, (vm, { op1: _debugInfo }) => {
+    let debuggerInfo = vm.constants.getValue<DebuggerInfo>(decodeHandle(_debugInfo));
+    let inspector = new ScopeInspector(vm.scope(), debuggerInfo);
+    callback(valueForRef(vm.getSelf()), (path) => valueForRef(inspector.get(path)));
+  });
+}
