+++ packages/@glimmer/runtime/lib/compiled/opcodes/expressions.ts
@@ -55,7 +55,7 @@ import { isIndexable } from '@glimmer/util/lib/collections';
 import { $v0 } from '@glimmer/vm/lib/registers';
 
 import { isCurriedType, resolveCurriedValue } from '../../curried-value';
-import { APPEND_OPCODES } from '../../opcodes';
+import type { AppendOpcodes } from '../../opcodes';
 import createCurryRef from '../../references/curry-value';
 import { reifyPositional } from '../../vm/arguments';
 import { createConcatRef } from '../expressions/concat';
@@ -69,82 +69,6 @@ import {
   CheckScopeBlock,
   CheckUndefinedReference,
 } from './-debug-strip';
-
-APPEND_OPCODES.add(VM_CURRY_OP, (vm, { op1: type, op2: _isStrict }) => {
-  let stack = vm.stack;
-
-  let definition = check(stack.pop(), CheckReference);
-  let capturedArgs = check(stack.pop(), CheckCapturedArguments);
-
-  let owner = vm.getOwner();
-  let resolver = vm.context.resolver;
-
-  let isStrict = false;
-
-  if (DEBUG) {
-    // strict check only happens in import.meta.env.DEV builds, no reason to load it otherwise
-    isStrict = vm.constants.getValue<boolean>(decodeHandle(_isStrict));
-  }
-
-  vm.loadValue(
-    $v0,
-    createCurryRef(type as CurriedType, definition, owner, capturedArgs, resolver, isStrict)
-  );
-});
-
-APPEND_OPCODES.add(VM_DYNAMIC_HELPER_OP, (vm) => {
-  let stack = vm.stack;
-  let ref = check(stack.pop(), CheckReference);
-  let args = check(stack.pop(), CheckArguments).capture();
-
-  let helperRef: Initializable<Reference>;
-  let initialOwner = vm.getOwner();
-
-  let helperInstanceRef = createComputeRef(() => {
-    if (helperRef !== undefined) {
-      destroy(helperRef);
-    }
-
-    let definition = valueForRef(ref);
-
-    if (isCurriedType(definition, CURRIED_HELPER)) {
-      let { definition: resolvedDef, owner, positional, named } = resolveCurriedValue(definition);
-
-      let helper = resolveHelper(resolvedDef, ref);
-
-      if (named !== undefined) {
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-        args.named = assign({}, ...named, args.named);
-      }
-
-      if (positional !== undefined) {
-        args.positional = positional.concat(args.positional) as CapturedPositionalArguments;
-      }
-
-      helperRef = helper(args, owner);
-
-      associateDestroyableChild(helperInstanceRef, helperRef);
-    } else if (isIndexable(definition)) {
-      let helper = resolveHelper(definition, ref);
-      helperRef = helper(args, initialOwner);
-
-      if (_hasDestroyableChildren(helperRef)) {
-        associateDestroyableChild(helperInstanceRef, helperRef);
-      }
-    } else {
-      helperRef = UNDEFINED_REFERENCE;
-    }
-  });
-
-  let helperValueRef = createComputeRef(() => {
-    valueForRef(helperInstanceRef);
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion -- @fixme
-    return valueForRef(helperRef!);
-  });
-
-  vm.associateDestroyable(helperInstanceRef);
-  vm.loadValue($v0, helperValueRef);
-});
 
 function resolveHelper(definition: HelperDefinitionState, ref: Reference): Helper {
   let managerOrHelper = getInternalHelperManager(definition, true);
@@ -172,72 +96,6 @@ function resolveHelper(definition: HelperDefinitionState, ref: Reference): Helpe
   return helper;
 }
 
-APPEND_OPCODES.add(VM_HELPER_OP, (vm, { op1: handle }) => {
-  let stack = vm.stack;
-  let helper = check(vm.constants.getValue(handle), CheckHelper);
-  let args = check(stack.pop(), CheckArguments);
-  let value = helper(args.capture(), vm.getOwner(), vm.dynamicScope());
-
-  if (_hasDestroyableChildren(value)) {
-    vm.associateDestroyable(value);
-  }
-
-  vm.loadValue($v0, value);
-});
-
-APPEND_OPCODES.add(VM_GET_VARIABLE_OP, (vm, { op1: symbol }) => {
-  let expr = vm.referenceForSymbol(symbol);
-
-  vm.stack.push(expr);
-});
-
-APPEND_OPCODES.add(VM_SET_VARIABLE_OP, (vm, { op1: symbol }) => {
-  let expr = check(vm.stack.pop(), CheckReference);
-  vm.scope().bindSymbol(symbol, expr);
-});
-
-APPEND_OPCODES.add(VM_SET_BLOCK_OP, (vm, { op1: symbol }) => {
-  let handle = check(vm.stack.pop(), CheckCompilableBlock);
-  let scope = check(vm.stack.pop(), CheckScope);
-  let table = check(vm.stack.pop(), CheckBlockSymbolTable);
-
-  vm.scope().bindBlock(symbol, [handle, scope, table]);
-});
-
-APPEND_OPCODES.add(VM_ROOT_SCOPE_OP, (vm, { op1: size }) => {
-  vm.pushRootScope(size, vm.getOwner());
-});
-
-APPEND_OPCODES.add(VM_GET_PROPERTY_OP, (vm, { op1: _key }) => {
-  let key = vm.constants.getValue<string>(_key);
-  let expr = check(vm.stack.pop(), CheckReference);
-  vm.stack.push(childRefFor(expr, key));
-});
-
-APPEND_OPCODES.add(VM_GET_BLOCK_OP, (vm, { op1: _block }) => {
-  let { stack } = vm;
-  let block = vm.scope().getBlock(_block);
-
-  stack.push(block);
-});
-
-APPEND_OPCODES.add(VM_SPREAD_BLOCK_OP, (vm) => {
-  let { stack } = vm;
-  let block = check(stack.pop(), CheckNullable(CheckOr(CheckScopeBlock, CheckUndefinedReference)));
-
-  if (block && !isUndefinedReference(block)) {
-    let [handleOrCompilable, scope, table] = block;
-
-    stack.push(table);
-    stack.push(scope);
-    stack.push(handleOrCompilable);
-  } else {
-    stack.push(null);
-    stack.push(null);
-    stack.push(null);
-  }
-});
-
 function isUndefinedReference(input: ScopeBlock | Reference): input is Reference {
   assert(
     Array.isArray(input) || input === UNDEFINED_REFERENCE,
@@ -246,88 +104,238 @@ function isUndefinedReference(input: ScopeBlock | Reference): input is Reference
   return input === UNDEFINED_REFERENCE;
 }
 
-APPEND_OPCODES.add(VM_HAS_BLOCK_OP, (vm) => {
-  let { stack } = vm;
-  let block = check(stack.pop(), CheckNullable(CheckOr(CheckScopeBlock, CheckUndefinedReference)));
+export function defineExpressionOpcodes(APPEND_OPCODES: AppendOpcodes): void {
+  APPEND_OPCODES.add(VM_CURRY_OP, (vm, { op1: type, op2: _isStrict }) => {
+    let stack = vm.stack;
 
-  if (block && !isUndefinedReference(block)) {
-    stack.push(TRUE_REFERENCE);
-  } else {
-    stack.push(FALSE_REFERENCE);
-  }
-});
+    let definition = check(stack.pop(), CheckReference);
+    let capturedArgs = check(stack.pop(), CheckCapturedArguments);
 
-APPEND_OPCODES.add(VM_HAS_BLOCK_PARAMS_OP, (vm) => {
-  // FIXME(mmun): should only need to push the symbol table
-  let block = vm.stack.pop();
-  let scope = vm.stack.pop();
+    let owner = vm.getOwner();
+    let resolver = vm.context.resolver;
 
-  check(block, CheckMaybe(CheckOr(CheckHandle, CheckCompilableBlock)));
-  check(scope, CheckMaybe(CheckScope));
-  let table = check(vm.stack.pop(), CheckMaybe(CheckBlockSymbolTable));
+    let isStrict = false;
 
-  let hasBlockParams = table && table.parameters.length;
-  vm.stack.push(hasBlockParams ? TRUE_REFERENCE : FALSE_REFERENCE);
-});
+    if (DEBUG) {
+      // strict check only happens in import.meta.env.DEV builds, no reason to load it otherwise
+      isStrict = vm.constants.getValue<boolean>(decodeHandle(_isStrict));
+    }
 
-APPEND_OPCODES.add(VM_CONCAT_OP, (vm, { op1: count }) => {
-  let out = new Array<Reference>(count);
+    vm.loadValue(
+      $v0,
+      createCurryRef(type as CurriedType, definition, owner, capturedArgs, resolver, isStrict)
+    );
+  });
 
-  for (let i = count; i > 0; i--) {
-    let offset = i - 1;
-    out[offset] = check(vm.stack.pop(), CheckReference);
-  }
+  APPEND_OPCODES.add(VM_DYNAMIC_HELPER_OP, (vm) => {
+    let stack = vm.stack;
+    let ref = check(stack.pop(), CheckReference);
+    let args = check(stack.pop(), CheckArguments).capture();
 
-  vm.stack.push(createConcatRef(out));
-});
+    let helperRef: Initializable<Reference>;
+    let initialOwner = vm.getOwner();
 
-APPEND_OPCODES.add(VM_IF_INLINE_OP, (vm) => {
-  let condition = check(vm.stack.pop(), CheckReference);
-  let truthy = check(vm.stack.pop(), CheckReference);
-  let falsy = check(vm.stack.pop(), CheckReference);
-
-  vm.stack.push(
-    createComputeRef(() => {
-      if (toBool(valueForRef(condition))) {
-        return valueForRef(truthy);
-      } else {
-        return valueForRef(falsy);
+    let helperInstanceRef = createComputeRef(() => {
+      if (helperRef !== undefined) {
+        destroy(helperRef);
       }
-    })
-  );
-});
 
-APPEND_OPCODES.add(VM_NOT_OP, (vm) => {
-  let ref = check(vm.stack.pop(), CheckReference);
+      let definition = valueForRef(ref);
 
-  vm.stack.push(
-    createComputeRef(() => {
-      return !toBool(valueForRef(ref));
-    })
-  );
-});
+      if (isCurriedType(definition, CURRIED_HELPER)) {
+        let { definition: resolvedDef, owner, positional, named } = resolveCurriedValue(definition);
 
-APPEND_OPCODES.add(VM_GET_DYNAMIC_VAR_OP, (vm) => {
-  let scope = vm.dynamicScope();
-  let stack = vm.stack;
-  let nameRef = check(stack.pop(), CheckReference);
+        let helper = resolveHelper(resolvedDef, ref);
 
-  stack.push(
-    createComputeRef(() => {
-      let name = String(valueForRef(nameRef));
-      return valueForRef(scope.get(name));
-    })
-  );
-});
+        if (named !== undefined) {
+          // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+          args.named = assign({}, ...named, args.named);
+        }
 
-APPEND_OPCODES.add(VM_LOG_OP, (vm) => {
-  let { positional } = check(vm.stack.pop(), CheckArguments).capture();
+        if (positional !== undefined) {
+          args.positional = positional.concat(args.positional) as CapturedPositionalArguments;
+        }
 
-  vm.loadValue(
-    $v0,
-    createComputeRef(() => {
-      // eslint-disable-next-line no-console
-      console.log(...reifyPositional(positional));
-    })
-  );
-});
+        helperRef = helper(args, owner);
+
+        associateDestroyableChild(helperInstanceRef, helperRef);
+      } else if (isIndexable(definition)) {
+        let helper = resolveHelper(definition, ref);
+        helperRef = helper(args, initialOwner);
+
+        if (_hasDestroyableChildren(helperRef)) {
+          associateDestroyableChild(helperInstanceRef, helperRef);
+        }
+      } else {
+        helperRef = UNDEFINED_REFERENCE;
+      }
+    });
+
+    let helperValueRef = createComputeRef(() => {
+      valueForRef(helperInstanceRef);
+      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion -- @fixme
+      return valueForRef(helperRef!);
+    });
+
+    vm.associateDestroyable(helperInstanceRef);
+    vm.loadValue($v0, helperValueRef);
+  });
+
+  APPEND_OPCODES.add(VM_HELPER_OP, (vm, { op1: handle }) => {
+    let stack = vm.stack;
+    let helper = check(vm.constants.getValue(handle), CheckHelper);
+    let args = check(stack.pop(), CheckArguments);
+    let value = helper(args.capture(), vm.getOwner(), vm.dynamicScope());
+
+    if (_hasDestroyableChildren(value)) {
+      vm.associateDestroyable(value);
+    }
+
+    vm.loadValue($v0, value);
+  });
+
+  APPEND_OPCODES.add(VM_GET_VARIABLE_OP, (vm, { op1: symbol }) => {
+    let expr = vm.referenceForSymbol(symbol);
+
+    vm.stack.push(expr);
+  });
+
+  APPEND_OPCODES.add(VM_SET_VARIABLE_OP, (vm, { op1: symbol }) => {
+    let expr = check(vm.stack.pop(), CheckReference);
+    vm.scope().bindSymbol(symbol, expr);
+  });
+
+  APPEND_OPCODES.add(VM_SET_BLOCK_OP, (vm, { op1: symbol }) => {
+    let handle = check(vm.stack.pop(), CheckCompilableBlock);
+    let scope = check(vm.stack.pop(), CheckScope);
+    let table = check(vm.stack.pop(), CheckBlockSymbolTable);
+
+    vm.scope().bindBlock(symbol, [handle, scope, table]);
+  });
+
+  APPEND_OPCODES.add(VM_ROOT_SCOPE_OP, (vm, { op1: size }) => {
+    vm.pushRootScope(size, vm.getOwner());
+  });
+
+  APPEND_OPCODES.add(VM_GET_PROPERTY_OP, (vm, { op1: _key }) => {
+    let key = vm.constants.getValue<string>(_key);
+    let expr = check(vm.stack.pop(), CheckReference);
+    vm.stack.push(childRefFor(expr, key));
+  });
+
+  APPEND_OPCODES.add(VM_GET_BLOCK_OP, (vm, { op1: _block }) => {
+    let { stack } = vm;
+    let block = vm.scope().getBlock(_block);
+
+    stack.push(block);
+  });
+
+  APPEND_OPCODES.add(VM_SPREAD_BLOCK_OP, (vm) => {
+    let { stack } = vm;
+    let block = check(
+      stack.pop(),
+      CheckNullable(CheckOr(CheckScopeBlock, CheckUndefinedReference))
+    );
+
+    if (block && !isUndefinedReference(block)) {
+      let [handleOrCompilable, scope, table] = block;
+
+      stack.push(table);
+      stack.push(scope);
+      stack.push(handleOrCompilable);
+    } else {
+      stack.push(null);
+      stack.push(null);
+      stack.push(null);
+    }
+  });
+
+  APPEND_OPCODES.add(VM_HAS_BLOCK_OP, (vm) => {
+    let { stack } = vm;
+    let block = check(
+      stack.pop(),
+      CheckNullable(CheckOr(CheckScopeBlock, CheckUndefinedReference))
+    );
+
+    if (block && !isUndefinedReference(block)) {
+      stack.push(TRUE_REFERENCE);
+    } else {
+      stack.push(FALSE_REFERENCE);
+    }
+  });
+
+  APPEND_OPCODES.add(VM_HAS_BLOCK_PARAMS_OP, (vm) => {
+    // FIXME(mmun): should only need to push the symbol table
+    let block = vm.stack.pop();
+    let scope = vm.stack.pop();
+
+    check(block, CheckMaybe(CheckOr(CheckHandle, CheckCompilableBlock)));
+    check(scope, CheckMaybe(CheckScope));
+    let table = check(vm.stack.pop(), CheckMaybe(CheckBlockSymbolTable));
+
+    let hasBlockParams = table && table.parameters.length;
+    vm.stack.push(hasBlockParams ? TRUE_REFERENCE : FALSE_REFERENCE);
+  });
+
+  APPEND_OPCODES.add(VM_CONCAT_OP, (vm, { op1: count }) => {
+    let out = new Array<Reference>(count);
+
+    for (let i = count; i > 0; i--) {
+      let offset = i - 1;
+      out[offset] = check(vm.stack.pop(), CheckReference);
+    }
+
+    vm.stack.push(createConcatRef(out));
+  });
+
+  APPEND_OPCODES.add(VM_IF_INLINE_OP, (vm) => {
+    let condition = check(vm.stack.pop(), CheckReference);
+    let truthy = check(vm.stack.pop(), CheckReference);
+    let falsy = check(vm.stack.pop(), CheckReference);
+
+    vm.stack.push(
+      createComputeRef(() => {
+        if (toBool(valueForRef(condition))) {
+          return valueForRef(truthy);
+        } else {
+          return valueForRef(falsy);
+        }
+      })
+    );
+  });
+
+  APPEND_OPCODES.add(VM_NOT_OP, (vm) => {
+    let ref = check(vm.stack.pop(), CheckReference);
+
+    vm.stack.push(
+      createComputeRef(() => {
+        return !toBool(valueForRef(ref));
+      })
+    );
+  });
+
+  APPEND_OPCODES.add(VM_GET_DYNAMIC_VAR_OP, (vm) => {
+    let scope = vm.dynamicScope();
+    let stack = vm.stack;
+    let nameRef = check(stack.pop(), CheckReference);
+
+    stack.push(
+      createComputeRef(() => {
+        let name = String(valueForRef(nameRef));
+        return valueForRef(scope.get(name));
+      })
+    );
+  });
+
+  APPEND_OPCODES.add(VM_LOG_OP, (vm) => {
+    let { positional } = check(vm.stack.pop(), CheckArguments).capture();
+
+    vm.loadValue(
+      $v0,
+      createComputeRef(() => {
+        // eslint-disable-next-line no-console
+        console.log(...reifyPositional(positional));
+      })
+    );
+  });
+}
