+++ packages/@glimmer/runtime/lib/debug-render-tree-register.ts
@@ -1,0 +1,4 @@
+import DebugRenderTreeImpl from './debug-render-tree';
+import { registerDebugRenderTreeFactory } from './environment';
+
+registerDebugRenderTreeFactory(() => new DebugRenderTreeImpl());
