+++ tests/node-vitest/tree-shakability.test.js
@@ -76,6 +76,7 @@ it('[dev] has expected tree-shakable entrypoints', async () => {
       "ember-source/@ember/template-compiler/lib/plugins/transform-wrap-mount-and-outlet.js",
       "ember-source/@ember/template-compiler/lib/plugins/utils.js",
       "ember-source/@ember/template-compiler/lib/system/calculate-location-display.js",
+      "ember-source/@ember/template-factory/index.js",
       "ember-source/@ember/template/index.js",
       "ember-source/@ember/test/index.js",
       "ember-source/@ember/utils/lib/is-equal.js",
@@ -86,6 +87,7 @@ it('[dev] has expected tree-shakable entrypoints', async () => {
       "ember-source/@glimmer/env/index.js",
       "ember-source/@glimmer/global-context/index.js",
       "ember-source/@glimmer/node/index.js",
+      "ember-source/@glimmer/opcode-compiler/index.js",
       "ember-source/@glimmer/owner/index.js",
       "ember-source/@glimmer/util/index.js",
       "ember-source/@glimmer/vm/index.js",
@@ -107,6 +109,9 @@ it('[dev] has expected tree-shakable entrypoints', async () => {
       "ember-source/@ember/-internals/deprecations/index.js",
       "ember-source/@ember/-internals/environment/index.js",
       "ember-source/@ember/-internals/glimmer/index.js",
+      "ember-source/@ember/-internals/glimmer/lib/register-classic-helper.js",
+      "ember-source/@ember/-internals/glimmer/lib/register-curly-component.js",
+      "ember-source/@ember/-internals/glimmer/lib/syntax/register-routing-keywords.js",
       "ember-source/@ember/-internals/meta/index.js",
       "ember-source/@ember/-internals/meta/lib/meta.js",
       "ember-source/@ember/-internals/metal/index.js",
@@ -152,6 +157,7 @@ it('[dev] has expected tree-shakable entrypoints', async () => {
       "ember-source/@ember/enumerable/mutable.js",
       "ember-source/@ember/helper/index.js",
       "ember-source/@ember/instrumentation/index.js",
+      "ember-source/@ember/instrumentation/lib/internal-instrument.js",
       "ember-source/@ember/modifier/index.js",
       "ember-source/@ember/modifier/on.js",
       "ember-source/@ember/object/-internals.js",
@@ -202,7 +208,6 @@ it('[dev] has expected tree-shakable entrypoints', async () => {
       "ember-source/@ember/template-compiler/lib/runtime.js",
       "ember-source/@ember/template-compiler/lib/template.js",
       "ember-source/@ember/template-compiler/runtime.js",
-      "ember-source/@ember/template-factory/index.js",
       "ember-source/@ember/test/adapter.js",
       "ember-source/@ember/utils/index.js",
       "ember-source/@ember/utils/lib/compare.js",
@@ -211,10 +216,10 @@ it('[dev] has expected tree-shakable entrypoints', async () => {
       "ember-source/@ember/utils/lib/is_present.js",
       "ember-source/@ember/utils/lib/type-of.js",
       "ember-source/@glimmer/manager/index.js",
-      "ember-source/@glimmer/opcode-compiler/index.js",
       "ember-source/@glimmer/program/index.js",
       "ember-source/@glimmer/reference/index.js",
       "ember-source/@glimmer/runtime/index.js",
+      "ember-source/@glimmer/runtime/lib/debug-render-tree-register.js",
       "ember-source/@glimmer/tracking/index.js",
       "ember-source/@glimmer/tracking/primitives/cache/index.js",
       "ember-source/@glimmer/validator/index.js",
@@ -284,6 +289,7 @@ it('[prod] has expected tree-shakable entrypoints', async () => {
       "ember-source/@ember/template-compiler/lib/plugins/transform-wrap-mount-and-outlet.js",
       "ember-source/@ember/template-compiler/lib/plugins/utils.js",
       "ember-source/@ember/template-compiler/lib/system/calculate-location-display.js",
+      "ember-source/@ember/template-factory/index.js",
       "ember-source/@ember/template/index.js",
       "ember-source/@ember/test/index.js",
       "ember-source/@ember/utils/lib/is-equal.js",
@@ -294,6 +300,7 @@ it('[prod] has expected tree-shakable entrypoints', async () => {
       "ember-source/@glimmer/env/index.js",
       "ember-source/@glimmer/global-context/index.js",
       "ember-source/@glimmer/node/index.js",
+      "ember-source/@glimmer/opcode-compiler/index.js",
       "ember-source/@glimmer/owner/index.js",
       "ember-source/@glimmer/util/index.js",
       "ember-source/@glimmer/vm/index.js",
@@ -315,6 +322,9 @@ it('[prod] has expected tree-shakable entrypoints', async () => {
       "ember-source/@ember/-internals/deprecations/index.js",
       "ember-source/@ember/-internals/environment/index.js",
       "ember-source/@ember/-internals/glimmer/index.js",
+      "ember-source/@ember/-internals/glimmer/lib/register-classic-helper.js",
+      "ember-source/@ember/-internals/glimmer/lib/register-curly-component.js",
+      "ember-source/@ember/-internals/glimmer/lib/syntax/register-routing-keywords.js",
       "ember-source/@ember/-internals/metal/index.js",
       "ember-source/@ember/-internals/routing/index.js",
       "ember-source/@ember/-internals/runtime/index.js",
@@ -356,6 +366,7 @@ it('[prod] has expected tree-shakable entrypoints', async () => {
       "ember-source/@ember/enumerable/mutable.js",
       "ember-source/@ember/helper/index.js",
       "ember-source/@ember/instrumentation/index.js",
+      "ember-source/@ember/instrumentation/lib/internal-instrument.js",
       "ember-source/@ember/modifier/index.js",
       "ember-source/@ember/modifier/on.js",
       "ember-source/@ember/object/-internals.js",
@@ -395,7 +406,6 @@ it('[prod] has expected tree-shakable entrypoints', async () => {
       "ember-source/@ember/template-compiler/lib/runtime.js",
       "ember-source/@ember/template-compiler/lib/template.js",
       "ember-source/@ember/template-compiler/runtime.js",
-      "ember-source/@ember/template-factory/index.js",
       "ember-source/@ember/test/adapter.js",
       "ember-source/@ember/utils/index.js",
       "ember-source/@ember/utils/lib/compare.js",
@@ -404,10 +414,10 @@ it('[prod] has expected tree-shakable entrypoints', async () => {
       "ember-source/@ember/utils/lib/is_present.js",
       "ember-source/@ember/utils/lib/type-of.js",
       "ember-source/@glimmer/manager/index.js",
-      "ember-source/@glimmer/opcode-compiler/index.js",
       "ember-source/@glimmer/program/index.js",
       "ember-source/@glimmer/reference/index.js",
       "ember-source/@glimmer/runtime/index.js",
+      "ember-source/@glimmer/runtime/lib/debug-render-tree-register.js",
       "ember-source/@glimmer/tracking/index.js",
       "ember-source/@glimmer/tracking/primitives/cache/index.js",
       "ember-source/@glimmer/validator/index.js",
